refactor(transfermarkt): split TransfermarktParser into focused parser classes
- Extract PlayerParser for player page parsing
- Extract SquadValueParser for squad value parsing
- Extract StandingParser for standing parsing
- Extract ClubParser for club page parsing
- Create thin TransfermarktParser facade

Fixes CLAUDE.md convention: 231 LOC -> 53 LOC (under 150 max)
